Исправляет функции проверки формы
